requirements.txt
Django
djangorestframework
python-dotenv
gunicorn
django-cors-headers
whitenoise
psycopg2-binary